ConfirmActionButton component

+++ StrayBase/Core/Profile/ProfileView.swift
@@ -58,57 +58,37 @@ struct ProfileView: View {
                 }
                 
                 Section(ProfileViewConsts.accountSectionTitle) {
-                    Button {
-                        showSignOutAlert = true
-                    } label: {
-                        SettingsRowView(
-                            imageName: ProfileViewConsts.signOutImageName,
-                            title: ProfileViewConsts.signOutTitle,
-                            tintColor: .red
-                        )
-                    }
-                    .alert(isPresented: $showSignOutAlert) {
-                        Alert(
-                            title: Text(ProfileViewConsts.signOutAlertTitle),
-                            message: Text(ProfileViewConsts.signOutAlertDescription),
-                            primaryButton: .destructive(Text(ProfileViewConsts.signOutAlertButton)) {
-                                Task {
-                                    do {
-                                        try await viewModel.signOut()
-                                    } catch {
-                                        print("DEBUG: Failed to sign out with error - \(error.localizedDescription)")
-                                    }
-                                }
-                            },
-                            secondaryButton: .cancel()
-                        )
-                    }
-                    
-                    Button {
-                        showDeleteAlert = true
-                    } label: {
-                        SettingsRowView(
-                            imageName: ProfileViewConsts.deleteAccountImageName,
-                            title: ProfileViewConsts.deleteAccountTitle,
-                            tintColor: .red
-                        )
-                    }
-                    .alert(isPresented: $showDeleteAlert) {
-                        Alert(
-                            title: Text(ProfileViewConsts.deleteAlertTitle),
-                            message: Text(ProfileViewConsts.deleteAlertDescription),
-                            primaryButton: .destructive(Text(ProfileViewConsts.deleteAlertButton)) {
-                                Task {
-                                    do {
-                                        try await viewModel.deleteAccount()
-                                    } catch {
-                                        print("DEBUG: Failed to delete account with error - \(error.localizedDescription)")
-                                    }
-                                }
-                            },
-                            secondaryButton: .cancel()
-                        )
-                    }
+                    ConfirmActionButton(viewModel: .init(
+                        imageName: ProfileViewConsts.signOutImageName,
+                        title: ProfileViewConsts.signOutTitle,
+                        tintColor: .red,
+                        alertTitle: ProfileViewConsts.signOutAlertTitle,
+                        alertMessage: ProfileViewConsts.signOutAlertDescription,
+                        alertConfirmButtonTitle: ProfileViewConsts.signOutAlertButton,
+                        action: {
+                            do {
+                                try await viewModel.signOut()
+                            } catch {
+                                print("DEBUG: Failed to sign out with error - \(error.localizedDescription)")
+                            }
+                        }
+                    ))
+
+                    ConfirmActionButton(viewModel: .init(
+                        imageName: ProfileViewConsts.deleteAccountImageName,
+                        title: ProfileViewConsts.deleteAccountTitle,
+                        tintColor: .red,
+                        alertTitle: ProfileViewConsts.deleteAlertTitle,
+                        alertMessage: ProfileViewConsts.deleteAlertDescription,
+                        alertConfirmButtonTitle: ProfileViewConsts.deleteAlertButton,
+                        action: {
+                            do {
+                                try await viewModel.deleteAccount()
+                            } catch {
+                                print("DEBUG: Failed to delete account with error - \(error.localizedDescription)")
+                            }
+                        }
+                    ))
                 }
             }
         }
